semana05/dia3/requirements.txt
django==3.2
djangorestframework
django-cors-headers